Move QuadratureRule
